Created Position2DImpl and Hitbox and updated all the code consequently

+++ src/main/java/gfight/common/Pair.java
@@ -5,26 +5,26 @@ package gfight.common;
  */
 public class Pair implements Position2D {
 
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
     /**
      * Constructor of the pair object.
      * @param x
      * @param y
      */
-    public Pair(final int x, final int y) {
+    public Pair(final double x, final double y) {
         this.x = x;
         this.y = y;
     }
 
     @Override
-    public final int getX() {
+    public final double getX() {
         return x;
     }
 
     @Override
-    public final int getY() {
+    public final double getY() {
         return y;
     }
